+++ packages/frontend/src/App.vue
@@ -4,8 +4,15 @@
 
 <script lang="ts">
 import { defineComponent } from "vue";
+import { useClient } from "villus";
 
 export default defineComponent({
   name: "App",
+  setup() {
+    // Villus GraphQL client
+    useClient({
+      url: "/graphql",
+    });
+  },
 });
 </script>